pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "human-multiview"
version = "0.1.0"
description = "Human-level 3D shape perception emerges from multi-view learning"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Tyler Bonnen", email = "bonnen@berkeley.edu"},
]
dependencies = [
    "torch>=2.0",
    "torchvision",
    "numpy",
    "scipy",
    "pandas",
    "matplotlib",
    "Pillow",
    "datasets",
    "scikit-image",
    "scikit-learn",
    "statsmodels",
    "tqdm",
]

[project.optional-dependencies]
notebooks = [
    "jupyter",
    "ipywidgets",
]

[tool.setuptools.packages.find]
include = ["human_multiview*"]
